requirements.txt
anyio==3.6.2
click==8.1.3
configparser==5.3.0
ecdsa==0.18.0
fastapi==0.95.2
greenlet==2.0.2
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
passlib==1.7.4
pdfkit==1.0.0
Pillow==9.5.0
psycopg2==2.9.6
pyasn1==0.5.0
pydantic==1.10.7
pypng==0.20220715.0
python-jose==3.3.0
python-multipart==0.0.6
qrcode==7.4.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.17
starlette==0.27.0
typing_extensions==4.5.0
uvicorn==0.22.0
